Add MSCC RNG algorithm
Add the MSCC RNG alongside the current functions.
The algorithm is described in my notes from the disassembly:
<https://tilde.town/~magical/chip/mscc-rng.txt>.

MS is about 1.5× slower than TW because the algorithm for blob movement
is more elaborate than TW and it makes more calls to the RNG function.
There's not much that we can do about that.

There are now two copies of moveBlob: moveBlobMS and moveBlobTW.
searchSeed picks the right movement function based on an "rng type" enum
passed into it. I tried passing moveBlob in as a function pointer
instead, but the overhead from making an indirect function call was too
great. Using an enum is much less costly and also allows us to print the
name of the ruleset when we find a successful seed.

Note that the order of the enum might be important - i got run times
that were 0.2 µs/seed slower the other way around. I'm not sure why;
perhaps it's affecting code layout? It could be measurement noise but it
seems pretty consistent. In any case, the difference is tiny.

Testing
-----

A Sickly Silver Moon used a memory editor to capture the state
of MSCC's RNG at the start of the level and after the blobs' first move.
I verified that starting the search program from the same seed results
in the same RNG state after the blobs move.

Here are the RNG values we tested:

| Initial    | After all blobs move once |
| ---------- | ------------------------- |
| 0xC6878D78 | 0xD0B9F3C6                |
| 0xB5FAC798 | 0xE9381162                |

We also dumped the map and checked that the positions of the blobs
match.