Improve verifyRoute error reporting a bit
if we run into a wall, don't report any other errors. when we run into a
wall we terminate the route (because we'll probably desync or go out of
bounds after that), so we definitely don't reach the exit and
probably haven't collected all the chips. so reporting either of those
errors after running into a wall isn't helpful.